+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change Log (vscode-powertools)
 
 
-## 0.10.0 (January 21st, 2019; config and tools)
+## 0.11.0 (January 21st, 2019; npm update & bug fixes)
+
+* bug fix: there way no auto creation of `.vscode-powertools` sub folder in home directory, when using [store](https://egodigital.github.io/vscode-powertools/api/classes/_stores_.userstore.html) in scripts, e.g. ... s. [issue #7](https://github.com/egodigital/vscode-powertools/issues/7)
+* updated the following [npm](https://www.npmjs.com/) modules:
+  * [moment](https://www.npmjs.com/package/moment) `^2.24.0`
+
+## 0.10.1 (January 21st, 2019; config and tools)
 
 * added `imports` setting, which can define one or more external setting files to import
 * added following functions for [Code Execution](https://github.com/egodigital/vscode-powertools/wiki/Tools#code-execution):